+++ xslang/sources/lib.rs
@@ -5,6 +5,7 @@
 
 pub mod codegen;
 pub mod compiler_core;
+pub mod gc;
 pub mod hir;
 pub mod mir;
 pub mod mono;